rename common.model to common.primitives
Primitives are the "build block" classes that can be used to build more complex data structures like the workbook classes
