--- common/src/main/kotlin/org/wycliffeassociates/otter/common/data/model/CollectionOrContent.kt
@@ -1,3 +1,0 @@
-package org.wycliffeassociates.otter.common.data.model
-
-interface CollectionOrContent
