+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/data/primitives/CollectionOrContent.kt
@@ -1,0 +1,3 @@
+package org.wycliffeassociates.otter.common.data.primitives
+
+interface CollectionOrContent
