+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/data/primitives/ContainerType.kt
@@ -1,4 +1,4 @@
-package org.wycliffeassociates.otter.common.data.model
+package org.wycliffeassociates.otter.common.data.primitives
 
 /**
  * Container type enum for [ResourceMetadata.type].
