+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/data/primitives/Content.kt
@@ -1,4 +1,4 @@
-package org.wycliffeassociates.otter.common.data.model
+package org.wycliffeassociates.otter.common.data.primitives
 
 data class Content(
     var sort: Int,
